requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
docopt==0.6.2
idna==3.11
numpy==2.3.4
oauthlib==3.3.1
pandas==2.3.3
pipreqs==0.4.13
python-dateutil==2.9.0.post0
python3-discogs-client==2.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
